Phase 5a: complete the querystore-only migration — remove the fallback + record-loading layer (#51)
Collapses chartsearchai onto querystore as the sole retrieval path and removes the
last in-process retrieval/serialization machinery. (Docs — README, the migration
plan, CLAUDE.md — follow in Phase 5b.)

- ChartBuildingStrategy.buildChart() -> unconditional queryStoreChartBuilder.build();
  the querystore.enabled=false full-chart fallback is gone. QueryStoreChartBuilder
  still degrades to an empty chart if querystore is unavailable, so there is no crash
  path.
- Deleted the record-loading layer: PatientRecordLoader + the 8 *TextSerializers
  (Clinical/Allergy/Condition/Diagnosis/MedicationDispense/Obs/Order/PatientProgram)
  + their 8 test classes. querystore produces the document text now.
- Relocated SerializedRecord from a nested class in PatientRecordLoader to a top-level
  serializer/SerializedRecord (the migration lynchpin); repointed all consumers.
- PatientChartSerializer: dropped the no-records serialize(patient) overload + its
  @Autowired PatientRecordLoader; keeps the with-records render path used by querystore.
- Removed the chartsearchai.querystore.enabled GP surface entirely: the GP +
  ChartSearchAiUtils.isQueryStoreEnabled + LlmInferenceService.resolveQueryStoreEnabled
  + the shouldRunWarmup queryStoreEnabled param + the answer-cache-key fold. querystore
  is a required module, so the toggle no longer had a meaning. Corrected the
  warmupEnabled GP description (warmup is viable in every querystore mode now).
- Trimmed dead code left by Phases 2-4: ChartSearchAiUtils.filterByCoherence/zScoreFloor,
  QueryPreprocessor.prepareEmbeddingInput/buildEmbeddingQuery, and the 14-constant
  embedding-scoring block in ChartSearchAiConstants (all zero-caller).
- Tests: rewrote LlmInferenceServiceWarmupTest for the 1-param gate; pruned the querystore
  dimension from the warmup integration test; removed resolveQueryStoreEnabled overrides
  from the async-grounding/citation-wiring stubs; fixed stale comments.

Verified: mvn -pl api test -> 467 run, 0 failures, 35 skipped (the context-sensitive
tests exercise the slimmed Spring context end-to-end). Clean omod carries no
record-loading classes and SerializedRecord at its new location. Standalone smoke
(agnes-adams): both modules start cleanly, grounded answer returned via the
querystore-only path (30 citations).
