+++ api/src/main/java/org/openmrs/module/chartsearchai/ChartSearchAiConstants.java
@@ -19,110 +19,6 @@ public class ChartSearchAiConstants {
 
 	public static final String GP_EMBEDDING_PRE_FILTER = "chartsearchai.embedding.preFilter";
 
-	public static final int ADAPTIVE_MIN_RECORDS = 2;
-
-	public static final double ABSOLUTE_SIMILARITY_FLOOR = 0.25;
-
-	/** Minimum z-score for the tight-cluster bypass of the zero-keyword
-	 * z-score gate. When the ratio floor produces a small cluster AND
-	 * the initial z-score exceeds this threshold AND the max semantic
-	 * score is comfortably above the absolute floor, the cluster is
-	 * considered validated and the z-score gate is bypassed. */
-	public static final double FLOOR_RESCUE_MIN_Z_SCORE = 2.0;
-
-	/** Minimum number of records required for the z-score gate to
-	 * activate. Below this threshold, the score distribution has too
-	 * few data points for the z-score to be statistically meaningful. */
-	public static final int MIN_RECORDS_FOR_Z_SCORE = 30;
-
-	/** Absolute coherence floor below which a candidate is a true outlier.
-	 * If a candidate flagged for removal by the coherence gap detector has
-	 * average pairwise coherence at or above this value, the cut is
-	 * suppressed — the candidate genuinely belongs to the same topic.
-	 * Prevents duplicate/near-duplicate embeddings (identical text → cosine
-	 * 1.0 between them) from inflating the coherence range and making a
-	 * same-topic record look like an outlier. Empirically: same-topic
-	 * candidates removed incorrectly have coherence ~0.91+, while true
-	 * cross-topic outliers have coherence ~0.49-. Value of 0.70 sits
-	 * well between these ranges. */
-	public static final double COHERENCE_SAME_TOPIC_FLOOR = 0.70;
-
-	/** Minimum cosine similarity between a keyword-matched record and
-	 * the semantic core (non-keyword records found by gap detection)
-	 * for the keyword match to be considered topically relevant.
-	 * Keyword reps below this threshold are incidental matches —
-	 * e.g. SpO2 matching "blood" in "blood oxygen saturation" when
-	 * the query is about blood problems. Value of 0.55 sits between
-	 * same-topic cross-type pairs (~0.65-0.80) and cross-topic pairs
-	 * (~0.40-0.50). */
-	public static final double SEMANTIC_CORE_MIN_COSINE = 0.55;
-
-	/** Fraction of the semantic core's minimum score used as the
-	 * relevance floor for expansion candidates. Semantic score
-	 * (query cosine) is the only signal where genuinely relevant
-	 * records consistently outscore template-similar noise — all
-	 * inter-record cosine metrics (max, avg, min to core) fail
-	 * because the embedding model conflates record-type similarity
-	 * with content similarity. Value of 0.80 means expansion
-	 * candidates must be within 20% of the core's relevance level. */
-	public static final double SEMANTIC_CORE_SCORE_RATIO = 0.80;
-
-
-	/** Fraction of max coherence used as the adaptive minimum gap for
-	 * coherence outlier detection. */
-	public static final double COHERENCE_ADAPTIVE_GAP_RATIO = 0.20;
-
-	/** Corpus size above which the selective-keyword rescue extends from
-	 * the absolute floor of {@link #ADAPTIVE_MIN_RECORDS} to a fractional
-	 * threshold. Smaller charts (typical synthetic eval datasets ~150
-	 * records) keep the absolute rule because their partial-keyword
-	 * fractions overlap with the eval-empty cases that the rescue must
-	 * not block. Larger charts (live patients with many concept entries)
-	 * have more headroom: a "selective" keyword set at 6/462 is 1.3 %,
-	 * which is meaningfully smaller than what eval cases produce, so a
-	 * larger absolute count still represents a rare keyword anchor. */
-	public static final int LARGE_CORPUS_SELECTIVE_RESCUE_MIN = 200;
-
-	/** Fraction-of-corpus threshold for the selective-keyword rescue on
-	 * large charts (corpus &gt; {@link #LARGE_CORPUS_SELECTIVE_RESCUE_MIN}).
-	 * Records matching a keyword that appears in &le; this fraction of
-	 * the corpus are considered selective enough to rescue from the
-	 * partial-match semantic floor. Calibrated to catch Richard-class
-	 * broad queries (musculoskeletal injuries: 6/462 = 1.3 %; tests
-	 * ordered: 24/462 = 5.2 %) without affecting smaller eval datasets
-	 * (which the size gate excludes). */
-	public static final double LARGE_CORPUS_SELECTIVE_KW_FRACTION = 0.06;
-
-	/** Maximum number of keyword terms a query can have for the concept-
-	 * similarity expansion to consider it. Longer queries are typically
-	 * multi-concept and unsafe to collapse onto a single chart concept's
-	 * tokens. */
-	public static final int CONCEPT_EXPANSION_MAX_KW_TERMS = 4;
-
-	/** Minimum length of a kwTerm for it to count as a vocabulary-overlap
-	 * anchor in the moderate-similarity guard. Two-letter tokens (e.g.
-	 * "is", "of") are too generic to discriminate. */
-	public static final int CONCEPT_EXPANSION_MIN_OVERLAP_LENGTH = 3;
-
-	/** Fraction of the corpus above which a concept token is considered
-	 * too common to anchor a keyword search after expansion. A token
-	 * appearing in &gt;20 % of records is generic (e.g. "method" on a
-	 * chart with many frequency obs). Threshold has a floor of 1 so
-	 * very small corpora do not lose every token. */
-	public static final double CONCEPT_EXPANSION_IDF_FRACTION = 0.20;
-
-	/** Reference candidate count for coherence gap ratio calibration.
-	 * {@link #COHERENCE_ADAPTIVE_GAP_RATIO} was calibrated for candidate
-	 * sets of this size. For smaller sets, the gap ratio is scaled up by
-	 * √((REFERENCE−1) / (n−1)) to account for the higher variance in
-	 * coherence estimates (fewer pairwise comparisons per candidate).
-	 * Value of 5 means: for n=4 the scale is modest (√(4/3)≈1.15);
-	 * for n≥5 the scale is 1.0 (no inflation). The coherence filter
-	 * is not called for n&lt;4 (too few pairwise comparisons). */
-	public static final int COHERENCE_REFERENCE_N = 5;
-
-	public static final String GP_QUERYSTORE_ENABLED = "chartsearchai.querystore.enabled";
-
 	/** Number of top results the querystore retrieval path requests for the
 	 *  focus-hint pass; tunes that path independently of any default. */
 	public static final String GP_QUERYSTORE_TOP_K = "chartsearchai.querystore.topK";
@@ -227,7 +123,7 @@ public class ChartSearchAiConstants {
 	 * The {@link #DEFAULT_GROUNDING_MIN_COSINE} of {@value #DEFAULT_GROUNDING_MIN_COSINE}
 	 * suits a wide-spread model like all-MiniLM-L6-v2 (chartsearchai's own default).
 	 * It is far too low for <em>e5</em> — the model querystore uses, which the
-	 * grounding verifier reuses when {@code chartsearchai.querystore.enabled=true}.
+	 * grounding verifier reuses (querystore is the only retrieval/embedding backend).
 	 * Measured e5 cosines (mean-pooled, no prefix): supported pairs ~0.83–0.96,
 	 * unrelated pairs ~0.75–0.80. So on an e5/querystore deployment set this to
 	 * <strong>~0.82</strong>; at {@value #DEFAULT_GROUNDING_MIN_COSINE} e5 marks
