--- api/src/main/java/org/openmrs/module/chartsearchai/serializer/ConditionTextSerializer.java
@@ -1,84 +1,0 @@
-/**
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
- * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
- *
- * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
- * graphic logo is a trademark of OpenMRS Inc.
- */
-package org.openmrs.module.chartsearchai.serializer;
-
-import org.openmrs.Condition;
-import org.openmrs.module.chartsearchai.util.ConceptNameUtil;
-import org.openmrs.module.chartsearchai.util.DateFormatUtil;
-import org.springframework.stereotype.Component;
-
-/**
- * Serializes a {@link Condition} into embedding-friendly text.
- *
- * <p>Example output: {@code "Condition: Type 2 Diabetes Mellitus. Status: ACTIVE. Verification:
- * CONFIRMED"}</p>
- *
- * <p><b>Date handling:</b> The onset date is excluded (it serves as the record
- * timestamp in {@link PatientRecordLoader}). The end/resolution date is included
- * as {@code "Resolved: <date>"} because it is clinically significant — it tells
- * the LLM whether a condition is still active.
- */
-@Component
-public class ConditionTextSerializer implements ClinicalTextSerializer<Condition> {
-
-	@Override
-	public String toText(Condition condition) {
-		StringBuilder sb = new StringBuilder();
-		String name = getConditionName(condition);
-		if (!name.isEmpty()) {
-			sb.append("Condition: ").append(name);
-		}
-		if (condition.getClinicalStatus() != null) {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Status: ").append(condition.getClinicalStatus());
-		}
-
-		if (condition.getVerificationStatus() != null) {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Verification: ").append(condition.getVerificationStatus());
-		}
-		if (condition.getAdditionalDetail() != null && !condition.getAdditionalDetail().trim().isEmpty()) {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Detail: ").append(condition.getAdditionalDetail().trim());
-		}
-		if (condition.getEndDate() != null) {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Resolved: ").append(DateFormatUtil.formatDate(condition.getEndDate()));
-			if (condition.getEndReason() != null && !condition.getEndReason().trim().isEmpty()) {
-				sb.append(" (").append(condition.getEndReason().trim()).append(")");
-			}
-		}
-
-		return sb.toString();
-	}
-
-	private String getConditionName(Condition condition) {
-		if (condition.getCondition() == null) {
-			return "";
-		}
-		if (condition.getCondition().getCoded() != null) {
-			return ConceptNameUtil.getName(condition.getCondition().getCoded());
-		}
-		if (condition.getCondition().getNonCoded() != null
-				&& !condition.getCondition().getNonCoded().trim().isEmpty()) {
-			return condition.getCondition().getNonCoded().trim();
-		}
-		return "";
-	}
-
-}
