--- api/src/main/java/org/openmrs/module/chartsearchai/serializer/PatientProgramTextSerializer.java
@@ -1,81 +1,0 @@
-/**
- * This Source Code Form is subject to the terms of the Mozilla Public License,
- * v. 2.0. If a copy of the MPL was not distributed with this file, You can
- * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
- * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
- *
- * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
- * graphic logo is a trademark of OpenMRS Inc.
- */
-package org.openmrs.module.chartsearchai.serializer;
-
-import org.openmrs.PatientProgram;
-import org.openmrs.PatientState;
-import org.openmrs.module.chartsearchai.util.ConceptNameUtil;
-import org.openmrs.module.chartsearchai.util.DateFormatUtil;
-import org.springframework.stereotype.Component;
-
-/**
- * Serializes a {@link PatientProgram} into embedding-friendly text.
- *
- * <p>Example output: {@code "Program: HIV Treatment. Enrolled: 2024-01-15. Status: Active.
- * Current state: On ART"}</p>
- *
- * <p><b>Date handling:</b> Both the enrollment date and completion date are
- * included in the text because they are clinically significant — they define
- * the program timeline. The enrollment date also serves as the record
- * timestamp in {@link PatientRecordLoader}.
- */
-@Component
-public class PatientProgramTextSerializer implements ClinicalTextSerializer<PatientProgram> {
-
-	@Override
-	public String toText(PatientProgram patientProgram) {
-		StringBuilder sb = new StringBuilder();
-
-		if (patientProgram.getProgram() != null && patientProgram.getProgram().getName() != null) {
-			sb.append("Program: ").append(patientProgram.getProgram().getName());
-			String description = patientProgram.getProgram().getDescription();
-			if (description != null && !description.trim().isEmpty()) {
-				sb.append(" (").append(description.trim()).append(")");
-			}
-		}
-
-		if (patientProgram.getDateEnrolled() != null) {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Enrolled: ").append(DateFormatUtil.formatDate(patientProgram.getDateEnrolled()));
-		}
-
-		if (patientProgram.getDateCompleted() != null) {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Completed: ").append(DateFormatUtil.formatDate(patientProgram.getDateCompleted()));
-		} else {
-			if (sb.length() > 0) {
-				sb.append(". ");
-			}
-			sb.append("Status: Active");
-		}
-
-		if (patientProgram.getOutcome() != null) {
-			String outcomeName = ConceptNameUtil.getName(patientProgram.getOutcome());
-			if (!outcomeName.isEmpty()) {
-				sb.append(". Outcome: ").append(outcomeName);
-			}
-		}
-
-		PatientState currentState = patientProgram.getCurrentState(null);
-		if (currentState != null && currentState.getState() != null
-				&& currentState.getState().getConcept() != null) {
-			String stateName = ConceptNameUtil.getName(currentState.getState().getConcept());
-			if (!stateName.isEmpty()) {
-				sb.append(". Current state: ").append(stateName);
-			}
-		}
-
-		return sb.toString();
-	}
-}
